Add .ruby-version file
This file originated from here I think: https://gist.github.com/fnichol/1912050

It's a good way to keep everyone on the project in sync with the same version of ruby. Also, you'll eventually have multiple versions of ruby installed with tools like rvm or chruby. This file can also help automatically switch you to that version of ruby when you enter the directory.