Add associations between types and pokemons
This requires two belongs_to associations for type_1 and type_2 respectively on the Pokemon model. This also requires two has_many associations on the Type model, one for each belongs_to on the Pokemon side.

As far as I know, there's no way to have a single association link to all Pokemon that have a certain type as either type_1 or type_2. To do that, we'll need a custom method. More on that later though. This is good for now.