Fixing the queries bugs, making views
It turns out I had a typo in the view name, and needed to have it link to the x_path, which I did by changing the name in the queries array.
I also finished a draft of the views for the first 3 queries. I'll probably go back and add more information to the views, but this is enough for now.