Query 4: Find all the pokemon that are bug or flying
Because I'm using Rails 4 for the moment, the ActiveRecord Query Interface does not support OR. So I had to build up this query in parts, finding all the type 1 bug, then the type 2 bug, then type 1 flying and then type 2 flying. While I could have created an array from the start and shoveled them in as I completed each query, I thought it would be clearer to assign them to variables.

As far as I know there are no side effects resulting from this commit.