Query 5a: Top 10 easiest to catch Pokemon
I had to choose a way to pick out the top 10 easiest to catch pokemon from a giant set of pokemon with a catch rate of 255. What else factors into the easiness of catching besides catch rate? If you have a heavy ball, weight does. So I took the list of catch rate 255 pokemon, ordered by weight descending and limited it to 10. Boom!

There are probably other ways to determine the top 10 easiest to catch pokemon, but weight seems to be the most immutable and is something you can order by, as opposed to saying water pokemon are easier to catch if you have a dive ball. And one might say the easiest to catch Pokemon is the one in front of you, but spawn rates and levels are determined by what area you're in.