Query 6: The easiest type of Pokemon to catch (on average)
I decided to go about this by finding the average catch rate for each type of pokemon, where a pokemon is of a given type if either their first or second type is that type. Eg. a Grass Fighting pokemon would count in both grass and fighting, and a bug grass would also count as a grass (and a bug). Yes, that means that dual typed pokemon would be double counted, but I would say those pokemon are equally of both types.

I also decided to show the whole sorted array of totals, so you can see what all the averages are.