Adding links from Bulbapedia to the DB
I feel like there's a lot of really cool information on Bulbapedia outside of the Pokedex entries here. So I added the links and images. Fortunately both followed a pattern and I was able to interpolate the names and ndex numbers to get the links and images. Though for Pokemon with multiple forms and pokedex entries, this might make things wonky.