Query 7 "Find a random pokemon with a good catch rate"
I decided to define "good" here as greater than or equal to 225 (out of 255). This is arbitrary. It's also super easy to change if you decide that something with a lower catch rate is still easy to catch. If you go to the view, you can refresh the page and get a new random easy to catch pokemon.