Query #11 List eact Type and the number of pokemon with that type as the primary type
In the process of writing this query, I discovered an off by one error in a previous query (easiest type to catch on average). The type id was not the index of the types array, but the index plus one. The error is fixed in both queries now, and it turns out the easiest type to catch is Poison and not Ground.