Query 12: 5 random male, female, and genderless Pokemon
Gender rate is an integer between -1 and 8. A value between 0 and 8 represents the ratio of male to female, where 0 is 100% male and 8 is 100% female. -1 is for genderless Pokemon.

In the process of working out this query, I've found an interesting bug, which is that pokemon with multiple forms have multiple pokedex entries (which I knew) and they'll show up multiple times in the sample, even though I should be getting unique entries. I think I want to keep the multiple forms in querydex, but I'll have to do something to account for them.