Adding Bulbapedia Images, Take 2
I was able to find the URLs via web scraping. I came across an interesting bug, where the names for the image pages weren't loading for Nidoran male/female due to the male and female signs in there names. I fixed this. For some reason the DB still won't seed, though, and I don't know why. So I commented out the code for now, and will work on it later.